requirements.txt
flake8==5.0.4
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
